add decimal for weight/height/girth and adds initial values for history

+++ src/MyFoodDoc.App.Application/Models/UserDto.cs
@@ -22,7 +22,7 @@ namespace MyFoodDoc.App.Application.Models
 
         public Gender? Gender { get; set; }
 
-        public int? Height { get; set; }
+        public decimal? Height { get; set; }
 
         public int? InsuranceId { get; set; }
