+++ src/MyFoodDoc.CMS/Models/VM/Patient.cs
@@ -12,11 +12,11 @@ namespace MyFoodDoc.CMS.Models.VM
         public string Email { get; set; }
         public int? InsuranceId { get; set; }
         public byte? Gender { get; set; }
-        public int Height { get; set; }
+        public decimal Height { get; set; }
         public DateTime? Birth { get; set; }
         public IList<HistoryBaseModel<decimal>> Weight { get; set; }
         //public IList<HistoryBaseModel<int>> BloodSugar { get; set; }
-        public IList<HistoryBaseModel<int>> AbdominalGirth { get; set; }
+        public IList<HistoryBaseModel<decimal>> AbdominalGirth { get; set; }
         public IList<string> Motivation { get; set; }
 
         public static Patient FromModel(PatientModel model)
@@ -30,7 +30,7 @@ namespace MyFoodDoc.CMS.Models.VM
                 InsuranceId = model.InsuranceId,
                 Height = model.Height ?? 0,
                 Gender = (byte?)model.Gender,
-                AbdominalGirth = model.AbdominalGirth?.Select(HistoryBaseModel<int>.FromModel).ToList(),
+                AbdominalGirth = model.AbdominalGirth?.Select(HistoryBaseModel<decimal>.FromModel).ToList(),
                 //BloodSugar = model.BloodSugar?.Select(HistoryBaseModel<int>.FromModel).ToList(),
                 Weight = model.Weight?.Select(HistoryBaseModel<decimal>.FromModel).ToList(),
                 Motivation = model.Motivation?.ToList()
